[BUGFIX] Avoid SQL error when module is opened on moved page
In case the backend module is opened on a page which was moved
within the current workspace there's a SQL error caused because
the processing for the relevant "Pagetree" runs empty.
Adding the current page to the processing can never break this part
because it will always be part of the result anyways.

Change-Id: I7e442392efc94d46baf23dbc17e57c1d097e9866
Resolves: #13635