[FEATURE] Integrate language handling in workspace module
Integrate language handling in workspace module. First a new
column "language" is integrated to the toolbar section that
shows the available sys_language records of the system. This
column allows to sort by the language uid as well.

Besides that a new virtual column "record status" is itegrated.
This is needed if e.g. a translated record and the default
parent record are new on the current workspace and only the
translated record shall be published - in this case the
localization might not be shown in the frontend since it just
cannot overlay the default record which is not available.
This behaviour is called "integrity check of records".

Change-Id: I89370281f6fdcd5d34e805b2d093a020fbdd4b2a
Resolves: #36790
Releases: 6.0
Reviewed-on: http://review.typo3.org/13902
Reviewed-by: Wouter Wolters
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader

Conflicts:
	Classes/ExtDirect/AbstractHandler.php
	Classes/ExtDirect/MassActionHandler.php
	Classes/ExtDirect/Server.php
	Classes/Service/Befunc.php
	Classes/Service/GridData.php
	Classes/Service/Workspaces.php
	Resources/Private/Language/locallang.xlf
	Resources/Public/JavaScript/Store/mainstore.js
	Resources/Public/JavaScript/configuration.js
	Resources/Public/JavaScript/grid.js
	Resources/Public/JavaScript/toolbar.js