[BUGFIX] Preview does not include L parameter on localization
The workspace module offers a way to preview accordant records
in the frontend. However, the L parameters is missing if dealing
with localized records.

The language needs to be determined from the version record, not
from the live record (which might be a placeholder as well).
Besides that, the L parameter needs to be provided as additional
GET parameter since it cannot be determined anymore later during
processing in t3lib_BEfunc::viewOnClick().

Change-Id: I97083cabce33c7f23a4f20687cfc81865cb8f68c
Fixes: #26612
Releases: 4.5, 4.6, 4.7, 6.0, 6.1
Reviewed-on: https://review.typo3.org/17277
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader